% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{commandlineArgs}
\alias{commandlineArgs}
\title{build commandline}
\usage{
commandlineArgs(commandName, args = list())
}
\description{
a helper function for build a command line string
for run docker commands.
}
