% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_functions.R
\name{create_container}
\alias{create_container}
\title{create_container}
\usage{
create_container(image_src = NULL, container_name = NULL,
  other_arguments = NULL, expose_port = NULL, port = NULL)
}
\description{
This function allows to create container based on an image
}
