% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_class.R
\name{docker}
\alias{docker}
\title{docker}
\description{
docker

docker
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-start}{\code{docker$start()}}
\item \href{#method-stop}{\code{docker$stop()}}
\item \href{#method-port}{\code{docker$port()}}
\item \href{#method-exec}{\code{docker$exec()}}
\item \href{#method-new}{\code{docker$new()}}
\item \href{#method-clone}{\code{docker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-start"></a>}}
\subsection{Method \code{start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\subsection{Method \code{stop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$stop(remove = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-port"></a>}}
\subsection{Method \code{port()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$port(filter_port = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exec"></a>}}
\subsection{Method \code{exec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$exec(command, sudo = F, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$new(
  image_src = NULL,
  container_name = NULL,
  other_arguments = NULL,
  expose_port = NULL,
  port = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{docker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
