% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_shiny_server.R
\docType{data}
\name{docker_shiny_server}
\alias{docker_shiny_server}
\title{docker_shiny_server}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
docker_shiny_server
}
\description{
docker_shiny_server
}
\keyword{datasets}
