% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium.R
\name{get_driver}
\alias{get_driver}
\title{get_driver}
\usage{
get_driver(port, ua = 1, cache = NULL)
}
\arguments{
\item{port}{specify which port you want to connect to}

\item{ua}{specify user agent id}

\item{cache}{specify cache}
}
\description{
get_driver
}
