% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{images}
\alias{images}
\title{docker images}
\usage{
images(all = FALSE, digests = FALSE, no_trunc = FALSE)
}
\arguments{
\item{all}{Show all images (default hides intermediate images)}

\item{digests}{Show digests}

\item{no_trunc}{Don’t truncate output}
}
\description{
List images
}
\details{
The default docker images will show all top level images,
their repository and tags, and their size.

Docker images have intermediate layers that increase reusability,
decrease disk usage, and speed up docker build by allowing each
step to be cached. These intermediate layers are not shown by
default.

The \code{SIZE} is the cumulative space taken up by the image and all its
parent images. This is also the disk space used by the contents of
the Tar file created when you docker save an image.

An image will be listed more than once if it has multiple repository
names or tags. This single image (identifiable by its matching \code{IMAGE
ID}) uses up the \code{SIZE} listed only once.
}
