% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_functions.R
\name{is_running}
\alias{is_running}
\title{is_running}
\usage{
is_running(name = NULL, image_src = NULL, quiet = T,
  return_logical = F)
}
\arguments{
\item{name}{A string providing the name of a container, whose state should be checked}

\item{image_src}{A string providing an image, from which derived containers should be checked}

\item{quiet}{Should the function return a logical in case a container matching the condition is running}
}
\value{
If a name is given as input, the function returns a message indicating whether a container named according to the input is running.

If an image is given as input, the fucntion returns a message indicating the number of containers built on the input images are running.

If quiet is False, the function will also return a logical vector indicating whether the condition is fulfilled.
}
\description{
This functions allows to check whether a container is running. One can either look by name or by image. If name and image are specified, name condition is prioritized.
}
