% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{ps}
\alias{ps}
\title{docker ps}
\usage{
ps(all = FALSE, no_trunc = FALSE, size = FALSE)
}
\arguments{
\item{all}{Show all containers (default shows just running)}

\item{no_trunc}{Don’t truncate output}

\item{size}{Display total file sizes}
}
\description{
List containers
}
\details{
list containers
}
