% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{docker run}
\usage{
run(
  container,
  commandline,
  workdir = "/",
  name = NULL,
  volume = NULL,
  tty = FALSE,
  lambda = NULL,
  framework = c("bash", ".netcore5", "mono")
)
}
\arguments{
\item{workdir}{Working directory inside the container}

\item{name}{Assign a name to the container}

\item{volume}{Bind mount a volume, see \link{volumeBind}.}

\item{lambda}{R function that running in \code{R#} environment.}
}
\description{
Run a command in a new container
}
\details{
The docker run command first creates a writeable container
layer over the specified image, and then starts it using the specified
command. That is, docker run is equivalent to the API
\code{/containers/create} then \code{/containers/(id)/start}. A stopped
container can be restarted with all its previous changes intact using
docker start. See docker ps -a to view a list of all containers.

The docker run command can be used in combination with docker commit to
change the command that a container runs. There is additional detailed
information about docker run in the Docker run reference.

For information on connecting a container to a network, see the
\code{Docker network overview}.
}
\seealso{
\link{volumeBind}
}
