% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sudo}
\alias{sudo}
\title{sudo}
\usage{
sudo(command, intern = F, ignore.stdout = F, ignore.stderr = T,
  env_var = T, cmd = F)
}
\arguments{
\item{command}{A bash command to execute as sudo}

\item{env_var}{Should the password be saved for further use}
}
\description{
This function executes the given command as sudo
}
