% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium.R
\name{switch_to_frame}
\alias{switch_to_frame}
\title{switch_to_frame}
\usage{
switch_to_frame(chrome, div_value = "", div_using = "css selector",
  frame_value = "", frame_using = "name")
}
\description{
switch_to_frame
}
