% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_functions.R
\name{view_container}
\alias{view_container}
\title{view_container}
\usage{
view_container(container_name, viewer = "vnc")
}
\arguments{
\item{container_name}{The name of a container to visualize}

\item{viewer}{the device to use to visualize the container. So far only vncviewer is supported.}
}
\description{
This function allows to visualize what is happening inside a container. Note that the container must expose the port 5900
}
