% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{volumeBind}
\alias{volumeBind}
\title{create volume binding}
\usage{
volumeBind(v)
}
\arguments{
\item{v}{a list object should in data structure like:
\code{list(anyname = list(host = ..., virtual = ...))}}
}
\description{
create volume binding
}
