% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker-package.r, R/zzz.R
\docType{package}
\name{docker}
\alias{docker}
\alias{docker-package}
\alias{docker}
\title{docker. Wraps Docker Python SDK.}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
docker
}
\description{
Allows accessing Docker SDK from R via the Docker \href{https://docker-py.readthedocs.io/en/stable/index.html}{Python SDK} using the reticulate package. This is a very thin wrapper and relies on the user to understand how the Python SDK works and how the \href{https://cran.r-project.org/package=reticulate}{reticulate} package is used to interface with python.

This is the docker instance corresponding to \href{https://docker-py.readthedocs.io/en/stable/index.html}{docker} module of the docker Python SDK.
}
\examples{
\dontrun{
library(docker)
client <- docker$from_env()
client$containers$run('alpine', 'echo "Hello World!"')
}

}
\keyword{datasets}
